refactor(session): extract the DAP mirror into src/session/mirror/
`exposeSession` / `unexposeSession` and their result types move to
`MirrorController`, which reaches the session record through the context's
`updateSession` rather than the store directly. Both result types are
re-exported from `session-manager-operations.ts`.
